// packages/browser/e2e-tests/cypress/lib/interfaces/dashboard/dashboardData.ts
export type TDashboardData = {
    URL: string;
    namePanel: string;
    labelOnePanelOne: string;
    labelTwoPanelOne: string;
    labelOnePanelTwo: string;
    labelTwoPanelTwo: string;
    labelOnePanelThree: string;
    labelTwoPanelThree: string;
    labelOnePanelFour: string;
    labelTwoPanelFour: string;
};